update cert store to use the static storage and fix validator bug in emulator
