+++ emulator/app/src/main.rs
@@ -715,7 +715,7 @@ fn run(cli: Emulator, capture_uart_output: bool) -> io::Result<Vec<u8>> {
             cli.i3c_port.unwrap(),
             i3c.get_dynamic_address().unwrap(),
             spdm_validator_tests,
-            Some(std::time::Duration::from_secs(4500)), // timeout in seconds
+            Some(std::time::Duration::from_secs(6000)), // timeout in seconds
         );
     }
