--- platforms/emulator/runtime/userspace/apps/user/src/spdm/cert_slot_mgr/mod.rs
@@ -1,6 +1,0 @@
-// Licensed under the Apache-2.0 license
-
-pub(crate) mod cert_store;
-pub(crate) mod device_cert;
-pub(crate) mod endorsement_cert;
-pub(crate) mod leaf_cert;
