+++ platforms/emulator/runtime/userspace/apps/user/src/spdm/cert_store/cert_chain/device.rs
@@ -1,9 +1,10 @@
 // Licensed under the Apache-2.0 license
 
-use crate::spdm::config::MAX_CERT_SIZE;
 use libapi_caliptra::certificate::CertContext;
 use spdm_lib::cert_store::{CertStoreError, CertStoreResult};
 use spdm_lib::protocol::AsymAlgo;
+
+const MAX_CERT_PORTION_SIZE: usize = 1024;
 
 pub enum DeviceCertIndex {
     IdevId, // Device Identity Certificate
@@ -48,18 +49,24 @@ impl DpeCertChain {
         if asym_algo != AsymAlgo::EccP384 {
             return Err(CertStoreError::UnsupportedAsymAlgo);
         }
+
+        if let Some(len) = self.cert_chain_len {
+            return Ok(len);
+        }
+
         let mut cert_chain_len = 0;
         let mut offset = self.cert_chain_offset().await;
+        let mut buf = [0u8; MAX_CERT_PORTION_SIZE];
+
         loop {
-            let size = self
-                .read_device_ecc_cert_chain(offset, &mut [0; MAX_CERT_SIZE])
-                .await?;
+            let size = self.read_device_ecc_cert_chain(offset, &mut buf).await?;
+            cert_chain_len += size;
             offset += size;
-            if size < MAX_CERT_SIZE {
-                cert_chain_len += offset;
+            if size < MAX_CERT_PORTION_SIZE {
                 break;
             }
         }
+
         self.cert_chain_len = Some(cert_chain_len);
         Ok(cert_chain_len)
     }
