+++ platforms/emulator/runtime/userspace/apps/user/src/spdm/cert_store/cert_chain/leaf.rs
@@ -1,6 +1,5 @@
 // Licensed under the Apache-2.0 license
 
-use crate::spdm::config::*;
 use embassy_sync::blocking_mutex::raw::CriticalSectionRawMutex;
 use embassy_sync::mutex::Mutex;
 use libapi_caliptra::certificate::CertContext;
@@ -8,6 +7,12 @@ use spdm_lib::cert_store::{CertStoreError, CertStoreResult};
 use spdm_lib::protocol::{AsymAlgo, ECC_P384_SIGNATURE_SIZE, SHA384_HASH_SIZE};
 
 const DPE_LEAF_CERT_SIZE: usize = 2048; // Size of the DPE leaf certificate buffer.
+
+pub const DPE_LEAF_CERT_LABEL: [u8; SHA384_HASH_SIZE] = [
+    0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f,
+    0x10, 0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19, 0x1a, 0x1b, 0x1c, 0x1d, 0x1e, 0x1f,
+    0x20, 0x21, 0x22, 0x23, 0x24, 0x25, 0x26, 0x27, 0x28, 0x29, 0x2a, 0x2b, 0x2c, 0x2d, 0x2e, 0x2f,
+];
 
 static SHARED_DPE_LEAF_CERT: Mutex<CriticalSectionRawMutex, DpeLeafCertBuf> =
     Mutex::new(DpeLeafCertBuf::new());
@@ -26,23 +31,23 @@ impl DpeLeafCert {
         dpe_leaf.reset();
     }
 
-    pub async fn size(&mut self, _asym_algo: AsymAlgo) -> CertStoreResult<usize> {
+    pub async fn size(&mut self, asym_algo: AsymAlgo) -> CertStoreResult<usize> {
         let mut dpe_leaf = SHARED_DPE_LEAF_CERT.lock().await;
         if dpe_leaf.size().is_none() {
-            dpe_leaf.fetch_cert(_asym_algo).await?;
+            dpe_leaf.fetch_cert(asym_algo).await?;
         }
         Ok(dpe_leaf.size().unwrap_or(0))
     }
 
     pub async fn read(
         &self,
-        _asym_algo: AsymAlgo,
+        asym_algo: AsymAlgo,
         offset: usize,
         buf: &mut [u8],
     ) -> CertStoreResult<usize> {
         let mut dpe_leaf = SHARED_DPE_LEAF_CERT.lock().await;
         if dpe_leaf.size().is_none() {
-            dpe_leaf.fetch_cert(_asym_algo).await?;
+            dpe_leaf.fetch_cert(asym_algo).await?;
         }
         dpe_leaf.read(offset, buf)
     }
@@ -89,6 +94,8 @@ impl DpeLeafCertBuf {
         if asym_algo != AsymAlgo::EccP384 {
             return Err(CertStoreError::UnsupportedAsymAlgo);
         }
+
+        self.reset();
 
         let mut cert_ctx = CertContext::new();
