+++ platforms/emulator/runtime/userspace/apps/user/src/spdm/endorsement_certs/slot0.rs
@@ -1,15 +1,5 @@
 // Licensed under the Apache-2.0 license
 
-use spdm_lib::protocol::SHA384_HASH_SIZE;
-
-pub const MAX_CERT_SIZE: usize = 1024;
-
-pub const DPE_LEAF_CERT_LABEL: [u8; SHA384_HASH_SIZE] = [
-    0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f,
-    0x10, 0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19, 0x1a, 0x1b, 0x1c, 0x1d, 0x1e, 0x1f,
-    0x20, 0x21, 0x22, 0x23, 0x24, 0x25, 0x26, 0x27, 0x28, 0x29, 0x2a, 0x2b, 0x2c, 0x2d, 0x2e, 0x2f,
-];
-
 pub static SLOT0_ECC_TEST_ROOT_CA_CERT_DER: [u8; 547] = [
     0x30, 0x82, 0x02, 0x1f, 0x30, 0x82, 0x01, 0xa5, 0x02, 0x01, 0x00, 0x30, 0x0a, 0x06, 0x08, 0x2a,
     0x86, 0x48, 0xce, 0x3d, 0x04, 0x03, 0x03, 0x30, 0x5e, 0x31, 0x1a, 0x30, 0x18, 0x06, 0x03, 0x55,
